@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
@@ -10,7 +10,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\TypeScriptBuilder\TypeScriptBuilder.csproj" />
+    <ProjectReference Include="..\TypeScriptBuilder\EntisCore.TypeScriptBuilder.csproj" />
   </ItemGroup>
 
 </Project>
